Added the loading mechanisim, for the gates, works. Disables win condition if the win was caused dsuring the loading. Fixed an issue in which the difficulty wasn't being set by the list box from the startgame form.
What to do now:
TO add a randomizer if the output is on after the loading is done.
Have a method in the END class checking if the on bool is true, which can be input in the LoadGates Method to check if the End is on. TO randomize If some buttons are on, to eventually loop until you get a button combination that makes the output not ON.

Graphics needs to be input, Text needs to be aligned on Background boxes. Wires can get a better look, maybe even image changing.